+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 
 # Changelog
 
+## v1.1.0 (_WIP_)
+
+* Allow caching costlocker responses with `--cache` option
+    * `--cache` - in verbose mode (`-vvv`) it also prints if costlocker is called or cached response is used
+    * `bin/clear-cache` - clear responses cache
+
+
 ## v1.0.1 (_2017-04-10_)
 
 * Add tags to summary report